introduce new MKL finding based on MKLConfig.cmake
* the MKLConfig.cmake is shipped with any recent (>=2021.3) Intel oneAPI installation
* it is searched for via CMake's find_package function
* basically works like link-line-advisor
* several new (advanced) config options are exposed to set e.g. static/dynamic linking
* the legacy behaviour is kept in to support older MKL versions, it SHOULD NOT be kept forever ...

Static and dynamic linking is supported on both Linux and Windows.
For some strange reason, static linking fails in the MSVC2019-Ifort2021 pipeline.
Therefore, static is the UNIX-default and dynamic the Windows-default.
Dynamic libraries are only copied over at the 'install' step (e.g. make install or cpack).
This avoids cmake warnings like "target is masked by files in bin" ...

The rpath in dynamic linking under Linux is working, but not correctly adapted (absolute oneAPI path remains there).
This seems to be a cmake bug ...